requirements.txt
telethon>=1.36.0
python-telegram-bot>=21.6
python-dotenv>=1.0.1
SQLAlchemy>=2.0.32
aiosqlite>=0.20.0
asyncpg>=0.29.0
APScheduler>=3.10.4
pandas>=2.2.2
openpyxl>=3.1.5
phonenumbers>=8.13.43
